refact(gaussian): Probability calculation module
A separate module has been created for the particle number detection and
threshold probability calculations.
